Working with strings again...

+++ src/engine/utils/IStringArray.h
@@ -8,6 +8,6 @@ typedef struct IStringArray IStringArray;
 IStringArray* IStringArray_Create();
 void IStringArray_Dispose(IStringArray* sa);
 void IStringArray_Add(IStringArray* sa, const char* str);
-size_t IStringArray_Length(IStringArray* sa);
+int64_t IStringArray_Length(IStringArray* sa);
 IString IStringArray_Get(IStringArray* sa, int32_t index);
 uint64_t IStringArray_GetRefs();
